[XABT] Refactor marshal methods and type map generation out of `GenerateJavaStubs`.
