+++ src/Xamarin.Android.Build.Tasks/Tasks/GenerateJavaStubs.cs
@@ -40,17 +40,8 @@ namespace Xamarin.Android.Tasks
 		[Required]
 		public string [] SupportedAbis { get; set; }
 
-		[Required]
-		public string TypemapOutputDirectory { get; set; }
-
-		[Required]
-		public bool GenerateNativeAssembly { get; set; }
-
 		public string IntermediateOutputDirectory { get; set; }
-		public bool LinkingEnabled { get; set; }
-		public bool HaveMultipleRIDs { get; set; }
 		public bool EnableMarshalMethods { get; set; }
-		public bool EnableManagedMarshalMethodsLookup { get; set; }
 
 		public bool Debug { get; set; }
 
@@ -63,19 +54,8 @@ namespace Xamarin.Android.Tasks
 
 		public string ApplicationJavaClass { get; set; }
 
-		public bool SkipJniAddNativeMethodRegistrationAttributeScan { get; set; }
-
-		public ITaskItem[] Environments { get; set; }
-
-		[Output]
-		public ITaskItem[] GeneratedBinaryTypeMaps { get; set; }
-
-		[Required]
-		public string AndroidRuntime { get; set; } = "";
-
 		public string CodeGenerationTarget { get; set; } = "";
 
-		AndroidRuntime androidRuntime;
 		JavaPeerStyle codeGenerationTarget;
 
 		internal const string AndroidSkipJavaStubGeneration = "AndroidSkipJavaStubGeneration";
@@ -83,7 +63,6 @@ namespace Xamarin.Android.Tasks
 		public override bool RunTask ()
 		{
 			try {
-				androidRuntime = MonoAndroidHelper.ParseAndroidRuntime (AndroidRuntime);
 				codeGenerationTarget = MonoAndroidHelper.ParseCodeGenerationTarget (CodeGenerationTarget);
 				bool useMarshalMethods = !Debug && EnableMarshalMethods;
 				Run (useMarshalMethods);
@@ -202,54 +181,6 @@ namespace Xamarin.Android.Tasks
 			}
 			JCWGenerator.EnsureAllArchitecturesAreIdentical (Log, nativeCodeGenStates);
 
-			if (useMarshalMethods) {
-				// We need to parse the environment files supplied by the user to see if they want to use broken exception transitions. This information is needed
-				// in order to properly generate wrapper methods in the marshal methods assembly rewriter.
-				// We don't care about those generated by us, since they won't contain the `XA_BROKEN_EXCEPTION_TRANSITIONS` variable we look for.
-				var environmentParser = new EnvironmentFilesParser ();
-				bool brokenExceptionTransitionsEnabled = environmentParser.AreBrokenExceptionTransitionsEnabled (Environments);
-
-				foreach (var kvp in nativeCodeGenStates) {
-					NativeCodeGenState state = kvp.Value;
-					if (!EnableManagedMarshalMethodsLookup) {
-						RewriteMarshalMethods (state, brokenExceptionTransitionsEnabled);
-						state.Classifier.AddSpecialCaseMethods ();
-					} else {
-						// We need to run `AddSpecialCaseMethods` before `RewriteMarshalMethods` so that we can see the special case
-						// methods (such as TypeManager.n_Activate_mm) when generating the managed lookup tables.
-						state.Classifier.AddSpecialCaseMethods ();
-						state.ManagedMarshalMethodsLookupInfo = new ManagedMarshalMethodsLookupInfo (Log);
-						RewriteMarshalMethods (state, brokenExceptionTransitionsEnabled);
-					}
-
-					Log.LogDebugMessage ($"[{state.TargetArch}] Number of generated marshal methods: {state.Classifier.MarshalMethods.Count}");
-					if (state.Classifier.RejectedMethodCount > 0) {
-						Log.LogWarning ($"[{state.TargetArch}] Number of methods in the project that will be registered dynamically: {state.Classifier.RejectedMethodCount}");
-					}
-
-					if (state.Classifier.WrappedMethodCount > 0) {
-						// TODO: change to LogWarning once the generator can output code which requires no non-blittable wrappers
-						Log.LogDebugMessage ($"[{state.TargetArch}] Number of methods in the project that need marshal method wrappers: {state.Classifier.WrappedMethodCount}");
-					}
-				}
-			}
-
-			bool typemapsAreAbiAgnostic = Debug && !GenerateNativeAssembly;
-			bool first = true;
-			foreach (var kvp in nativeCodeGenStates) {
-				if (!first && typemapsAreAbiAgnostic) {
-					Log.LogDebugMessage ("Typemaps: it's a debug build and type maps are ABI-agnostic, not processing more ABIs");
-					break;
-				}
-
-				NativeCodeGenState state = kvp.Value;
-				first = false;
-				WriteTypeMappings (state);
-			}
-
-			// Set for use by <GeneratePackageManagerJava/> task later
-			NativeCodeGenState.TemplateJniAddNativeMethodRegistrationAttributePresent = templateCodeGenState.JniAddNativeMethodRegistrationAttributePresent;
-
 			// Save NativeCodeGenState for later tasks
 			Log.LogDebugMessage ($"Saving {nameof (NativeCodeGenState)} to {nameof (NativeCodeGenStateRegisterTaskKey)}");
 			BuildEngine4.RegisterTaskObjectAssemblyLocal (MonoAndroidHelper.GetProjectBuildSpecificTaskObjectKey (NativeCodeGenStateRegisterTaskKey, WorkingDirectory, IntermediateOutputDirectory), nativeCodeGenStates, RegisteredTaskObjectLifetime.Build);
@@ -308,43 +239,6 @@ namespace Xamarin.Android.Tasks
 			}
 
 			return (allJavaTypes, javaTypesForJCW);
-		}
-
-		void RewriteMarshalMethods (NativeCodeGenState state, bool brokenExceptionTransitionsEnabled)
-		{
-			if (state.Classifier == null) {
-				return;
-			}
-
-			var rewriter = new MarshalMethodsAssemblyRewriter (Log, state.TargetArch, state.Classifier, state.Resolver, state.ManagedMarshalMethodsLookupInfo);
-			rewriter.Rewrite (brokenExceptionTransitionsEnabled);
-		}
-
-		void WriteTypeMappings (NativeCodeGenState state)
-		{
-			if (androidRuntime == Xamarin.Android.Tasks.AndroidRuntime.NativeAOT) {
-				// NativeAOT typemaps are generated in `Microsoft.Android.Sdk.ILLink.TypeMappingStep`
-				return;
-			}
-
-			Log.LogDebugMessage ($"Generating type maps for architecture '{state.TargetArch}'");
-			var tmg = new TypeMapGenerator (Log, state, androidRuntime);
-			if (!tmg.Generate (Debug, SkipJniAddNativeMethodRegistrationAttributeScan, TypemapOutputDirectory, GenerateNativeAssembly)) {
-				throw new XamarinAndroidException (4308, Properties.Resources.XA4308);
-			}
-
-			string abi = MonoAndroidHelper.ArchToAbi (state.TargetArch);
-			var items = new List<ITaskItem> ();
-			foreach (string file in tmg.GeneratedBinaryTypeMaps) {
-				var item = new TaskItem (file);
-				string fileName = Path.GetFileName (file);
-				item.SetMetadata ("DestinationSubPath", $"{abi}/{fileName}");
-				item.SetMetadata ("DestinationSubDirectory", $"{abi}/");
-				item.SetMetadata ("Abi", abi);
-				items.Add (item);
-			}
-
-			GeneratedBinaryTypeMaps = items.ToArray ();
 		}
 	}
 }
